arregla y mejora autocompletaciones de victima y victimacolectiva

+++ lib/sivel2_gen/concerns/controllers/gruposper_controller.rb
@@ -11,29 +11,46 @@ module Sivel2Gen
           include Msip::Concerns::Controllers::GruposperController
 
           def remplazar_antes_destruir_gp
+            true
           end 
 
           def remplazar_antes_salvar_vc
-          end 
+            true
+          end
+
+          def atributos_form
+            atributos_show - [:id, :caso_ids]
+          end
 
           # Busca y lista grupo(s) de persona
           def remplazar
             @grupoper = Msip::Grupoper.find(params[:grupoper_id].to_i)
-            @victimacolectiva = Sivel2Gen::Victimacolectiva.find(
-              params[:victima_idcolectiva].to_i)
-            grupoperant = @victimacolectiva.grupoper
-            @caso = @victimacolectiva.caso
+            @caso = Sivel2Gen::Caso.find(params[:caso_id].to_i)
             @caso.current_usuario = current_usuario
-            @victimacolectiva.grupoper = @grupoper
-            remplazar_antes_salvar_vc
-            @victimacolectiva.save!
-            if (grupoperant.nombre == 'N') ||
-              (grupoperant.nombre == '')
-              remplazar_antes_destruir_gp
-              grupoperant.destroy
+            if !remplazar_antes_salvar_vc
+              return
+            end
+            if Sivel2Gen::Victimacolectiva.where(grupoper_id: @grupoper.id).
+                where(caso_id: @caso.id).count == 0
+              @victimacolectiva = Sivel2Gen::Victimacolectiva.create(
+                grupoper_id: @grupoper.id, caso_id: @caso.id)
+              @victimacolectiva.save!
+              grupoperant = @victimacolectiva.grupoper
+              if (grupoperant.nombre == 'N') || (grupoperant.nombre == '')
+                remplazar_antes_destruir_gp
+                grupoperant.destroy
+              end
+            else
+              puts "Ya existe ese grupo de personas en el caso"
+              render json: "Ya existe ese grupo de personas en el caso",
+                status: :unprocessable_entity
+              return 
             end
             respond_to do |format|
-              format.html { render('remplazar', layout: false) }
+              format.html {
+                render('/msip/gruposper/remplazar', layout: false) 
+                return
+              }
             end
           end
